Completely rethink how models are generated
Use structs instead of traits, and those structs are just dumb stack-based wrappers for a json reference. This allows us to parse lazily, and avoid (hopefully) tons of unnecessary memory allocations for all the data we're not using. This is essentially a sparse view on the data. After parsing this will be slightly slower since it's lazy, but I don't think it'll be much overhead at all for rust!